ft_printf

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kato <kato@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/03 11:36:04 by kato              #+#    #+#             */
-/*   Updated: 2025/05/04 18:19:31 by kato             ###   ########.fr       */
+/*   Updated: 2025/05/04 19:09:22 by kato             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_putstr(char *str);
 int	ft_putnbr(int n);
 int	ft_putptr(unsigned long ptr);
 int	ft_putnbr_unsigned(unsigned int n);
-int	ft_putnbr_hex(unsigned int n, char format);
+int	ft_putnbr_hex(unsigned long n, char format);
 int	ft_printf(const char *format, ...);
 
 #endif
